requirements.txt
# FastAPI y servidor
fastapi>=0.115.12
fastapi-cli==0.0.7
#form data
python-multipart==0.0.20
#dovenv
python-dotenv>=1.0.0
# conexión AWS
boto3==1.35.87

#pydantic
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4

#slug
python-slugify==8.0.4

#bcrypt
bcrypt==4.3.0

# jwt
python-jose[cryptography]==3.1.0

#base de datos
psycopg2-binary==2.9.9
sqlmodel==0.0.24
alembic>=1.13.1
